requirements.txt
django
psycopg2-binary
python-dateutil
sentry-sdk
uwsgi
PyYAML
django-dotenv
django-extensions
social-auth-app-django
rest-social-auth
oauth2client
django-cors-headers
djangorestframework
dry-rest-permissions
django-import-export
django-admin-sortable2
pillow
sendgrid
